+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - [Multiple] Fix ckan version command and update copyrights (#4347 by: HebaruSan)
 - [GUI] Suppress exceptions on setting SplitterDistance (#4350 by: HebaruSan)
 - [Multiple] Detect future registry versions (#4364 by: HebaruSan)
+- [Multiple] Show download errors popup for downloads from Contents tab (#4366 by: HebaruSan)
 
 ### Internal